requirements.txt
aws-cdk.aws-apigatewayv2-alpha
aws-cdk.aws-apigatewayv2-authorizers-alpha
aws-cdk.aws-apigatewayv2-integrations-alpha
aws-cdk.aws-lambda-python-alpha
aws-cdk-lib>=2.116.0,<3.0.0
watchtower
pydantic